RF: remove check for pre-existing contents
Pre-existing contents only useful when you can be confident of the
entire context of the code.  That's true if the code runs in a fresh
namespace each time, but not when the namespace fills over the page, as
it does with nbplot.  I guess we could check whether the relevent
configuration and all previous code in the same page are the same across
runs, but for now, just remove the check.