RF: refactor nbplots to use node container
The previous code built some ReST, then inserted these ReST lines into
the ReST parser.   This meant that there is nothing in the doctree to
tell us that the contents started life as an nbplot directive, and
therefore, we can't detect literal blocks of code that should become
Jupyter notebook code cells.

Use `nested_parse` instead, to build the nodes, and then enclose them in
a container with a suitable class label, to tell us when we've got
nbplot contents.