RF: refactor way of allowing references to plots
The original matplotlib code did something rather nasty, detaching the
reference from the html only etc block and attaching to the figure
within.  That doesn't work correctly for latex etc, but in any case, we
want to make the reference point to the beginning of the block.  To make
that work, make the nbplot node a reference target.