RF+API: explicit nodes for plot links; code links
Add explicit nodes for plot epilogue and code links.  This allows us to
skip these nodes for the Markdown builders.

As part of that refactoring, remove the source code from the ReST
template used to make plot links etc.  The source code gets built
separately.  This is a minor API change from previous nb2plots and the
matplotlib plot_directive, in that the template (which is
user-configurable) should no longer include the source code.  So, if
anyone has been configuring the template, they'll get doubled code
output.